#3: Area() returs with double.

+++ 02Day/01AreaOfThePlanes/IPlane.cs
@@ -3,6 +3,6 @@
     public interface IPlane
     {
         //It's not necessary the "public" keyword because the interface has only public items.
-        int Area();
+        double Area();
     }
 }
